+++ driver/eeprom.c
@@ -31,8 +31,8 @@ void EEPROM_ReadBuffer(uint16_t Address, void *pBuffer, uint8_t Size)
 
 	I2C_Write(0xA0);
 
-	I2C_Write((Address >> 8) & 0xFF);
-	I2C_Write((Address >> 0) & 0xFF);
+	I2C_Write((uint8_t)((Address >> 8) & 0xFF));
+	I2C_Write((uint8_t)((Address >> 0) & 0xFF));
 
 	I2C_Start();
 
@@ -59,8 +59,8 @@ void EEPROM_WriteBuffer(uint16_t Address, const void *pBuffer)
 	taskENTER_CRITICAL();
 	I2C_Start();
 	I2C_Write(0xA0);
-	I2C_Write((Address >> 8) & 0xFF);
-	I2C_Write((Address >> 0) & 0xFF);
+	I2C_Write((uint8_t)((Address >> 8) & 0xFF));
+	I2C_Write((uint8_t)((Address >> 0) & 0xFF));
 	I2C_WriteBuffer(pBuffer, 8);
 	I2C_Stop();
